Use the new logger PoC

+++ iceoryx_posh/source/popo/ports/client_port_roudi.cpp
@@ -112,8 +112,8 @@ ClientPortRouDi::dispatchCaProMessageAndGetPossibleResponse(const capro::CaproMe
 void ClientPortRouDi::handleCaProProtocolViolation(const iox::capro::CaproMessageType messageType) noexcept
 {
     // this shouldn't be reached
-    LogFatal() << "CaPro Protocol Violation! Got '" << messageType << "' in `"
-               << getMembers()->m_connectionState.load(std::memory_order_relaxed) << "'";
+    LogFatal() << "CaPro Protocol Violation! Got '" << static_cast<uint64_t>(messageType) << "' in `"
+               << static_cast<uint64_t>(getMembers()->m_connectionState.load(std::memory_order_relaxed)) << "'";
     errorHandler(PoshError::POPO__CAPRO_PROTOCOL_ERROR, ErrorLevel::SEVERE);
 }
